Refactor big metric file into multiple files
