+++ main.go
@@ -62,7 +62,7 @@ func main() {
 	if whoopingAddress != nil && *whoopingAddress != "" {
 		logger := log.With("name", "whooper")
 		whoopDaemon := daemon.New(daemon.Configuration{
-			Reporter: defaultDaemonReporter(logger, prometheusReporter, "whopper"),
+			Reporter: prometheusReporter.Daemon(logger, "whopper"),
 			Interval: 1 * time.Second,
 			Tick: func() {
 				whooper.Whoop(*whoopingAddress, fmt.Sprintf("http://localhost%s", *listenAddress))
@@ -105,7 +105,7 @@ func main() {
 			With("port", port)
 		logger.Infof("Start checking address %s:%v", address, port)
 		tcpCheckerDaemon := daemon.New(daemon.Configuration{
-			Reporter: defaultDaemonReporter(logger, prometheusReporter, "tcpchecker"),
+			Reporter: prometheusReporter.Daemon(logger, "tcpchecker"),
 			Interval: 1 * time.Second,
 			Tick: func() {
 				tcpchecker.Check(name, address, int(port), tcpchecker.CompositeReporter(tcpchecker.LogReporter(logger), prometheusReporter.TcpChecker()))
@@ -149,10 +149,10 @@ func main() {
 		client := vici.NewClientConn(conn)
 		defer client.Close()
 		d := daemon.New(daemon.Configuration{
-			Reporter: defaultDaemonReporter(log.Base().With("name", "strongswan"), prometheusReporter, "strongswan"),
+			Reporter: prometheusReporter.Daemon(log.Base().With("name", "strongswan"), "strongswan"),
 			Interval: 2 * time.Second,
 			Tick: func() {
-				strongswan.Collect(client, prometheusReporter)
+				strongswan.Collect(client, prometheusReporter.StrongSwan())
 			},
 		})
 
@@ -178,24 +178,5 @@ func main() {
 		exitCode = 1
 	} else {
 		log.Info("exited due to a component shutting down")
-	}
-}
-
-func defaultDaemonReporter(logger log.Logger, prometheusReporter *metrics.PrometheusReporter, name string) *daemon.Reporter {
-	return &daemon.Reporter{
-		Started: func(d time.Duration) {
-			logger.With("state", "started").Infof("%s daemon started with interval %v", name, d)
-			prometheusReporter.Daemon.Started.WithLabelValues(name, d.String()).Inc()
-		},
-		Stopped: func() {
-			logger.With("state", "stopped").Infof("%s daemon stopped", name)
-			prometheusReporter.Daemon.Stopped.WithLabelValues(name).Inc()
-		},
-		Skipped: func() {
-			prometheusReporter.Daemon.Skipped.WithLabelValues(name).Inc()
-		},
-		Ticked: func() {
-			prometheusReporter.Daemon.Ticked.WithLabelValues(name).Inc()
-		},
 	}
 }
